fb: restore navbar compatibility with several xposed modules
I am tired of the complains....

Xposed use the emulator overwrite to enable/disable navbar...this results
into showing up navbar on reboot even if the user has set it to off in settings.

This fixes the issue and still keeps for emulator tests the basic overwrite functionallity
alive as long the provider value was not set.

We may should reflect that in settings as well.

Change-Id: Id60b45645d489c08539d56b8e05d135c36d38990

Patchset: 4
https://gerrit.slimroms.net/#/c/3098/