AppOps cm-10.2: Workaround gapps setupwizard fc
The only issue is that the AppOps from cm-10.2
fails to activate wifi when the users flash the
rom with a clean install. This causes the
gapps setupwizard to force close.

To prevent this issue for now, we set wifi to be activated
by default on a clean install of the rom.

Signed-off-by: Dragon7780 <Dragon7780@hotmail.nl>